Make the script work on Ubuntu Linux
The use of 'grep -v' was making a conflict with the use of 'bash -e'.
It looks like the return value logic of grep is differente on OS X and Linux.
Replaced by a awk call to count the number of entries not matching the pattern.

Added a .gitignore file for the repos which are pulled in.