Use a local m2repo for building the IDE
I've decided to do so because otherwise the Sbt artifacts built during the
script's execution would not be picked up when the
``org.scala-ide.sbt.full.library`` project is built. This is not ideal, but at
least it's easier to make sure that the Sbt artifacts packaged in the IDE are
indeed the ones built with this script (i.e., the artifacts are not downloaded
from the Typesafe repository).

As part of this commit, I've added three (3) new parameters needed to build
sbt from a given branch (this is needed for the IDE 2.1-M3 release).