Added step for building the worksheet
I had to add three new variables in order to have the
flexibility of building the worksheet when running this
script:

* $ECLIPSE_PLATFORM is needed to select the right
profile to pass to the Worksheet build.
* $BUILD_PLUGINS allows to build the Scala IDE alone,
which is useful for instance when using this script in a
Jenkins job for producing a nightly.
* WORKSHEET_BRANCH is the branch in the Worksheet
GH repository to use when to building the Worksheet.

Re #1001426