Clean up all local changes before building a project
This is done to avoid the script fails because of  some
local change. For instance, when setting the bundles'
version in the MANIFEST via the "set-versions" profile,
some local change is created (in the specific case, a
copy of the MANIFEST, i.e., MANIFEST.original). If we
don't clean up, next time the script is run it will fail.