Updates needed to build 4.0.0-b01
Adds support to specify a SBT version name (to be used in build toolchain and scala-ide).
Adds support to specify a version tag for the worksheet build.
Updates the worksheet eclipse profiles to the new name (from plugin-profiles).
Updates the scala profiles used for the worksheet build (from plugin-profiles).
Adds info about the 2.11.x ecosystem.