Modularizes building plugins, adds Scala-Search
All plugin building optional, default is just building the worksheet (as usual).
Passing a value to the branch for any plugin assumes this branch should be built.
If the branch is empty, the user iw prompted whether

Cleaned up the logic somewhat (branch existence is now tested even for non-interactive input,
plugin repos are not cloned if the plugin is not going to be built, branch asserts fail now before cloning).

This is very much still a rough cut.