Hide mvn output for checkAvailable
The output of the mvn run used to check if an artifact is available
in the m2 repos is not useful most of the time, and the possible failure
output can be misleading for people who don't know the internals of
the script.
The output is know put in target/log-extra.txt, and not display to the
user except in DEBUG mode.