Lift requirement to run build only with Java 6
The requirement existed because there was the believe that package
signing doesn't work with a Java version newer than 6. Fortunately I
couldn't find any issues with this. On Java 8 everything seems to work
as it should. I therefore lift the requirement and enforce that at least
Java 8 shall be used. Since we now depend on a sbt build we need lots of
perm memory for it and I couldn't find out how configure it for sbt in a
way that it considers it everywhere. Java 8 fortunately does not have
this problem since the `MaxPermSize` option no longer exists. Therefore
the enforcement to use Java 8.