Do no longer build zinc during the uber-build
It is really complicated to build zinc, which was part of sbt for most
of its live. Since zinc is now available as a standalone dependency,
which provides pre-built binaries, we can directly depend on the
binaries. All of the machinery that was needed to build zinc is
therefore no longer needed.

The changes that are necessary to depend on the pre-built zinc are part
of scala-ide and not of uber-build.